requirements.txt
fastapi
uvicorn
python-multipart
pypdf2
sentence-transformers
scikit-learn
bcrypt
requests
transformers
pillow